fix(pagination, differential): clean up pagination fixes, remove differential
