+++ package.json
@@ -67,7 +67,6 @@
     "metalsmith-build-date": "0.2.0",
     "metalsmith-collections": "0.7.0",
     "metalsmith-define": "2.0.1",
-    "metalsmith-differential": "git+https://github.com/mrajo/metalsmith-differential.git",
     "metalsmith-drafts": "0.0.1",
     "metalsmith-excerpts": "1.1.0",
     "metalsmith-fingerprint": "1.0.3",
